fix: Resolve client/server component boundary issues
- Simplify layout.tsx to avoid server component issues with Material-UI
- Move ThemeProvider and Material-UI setup to client component (page.tsx)
- Add 'use client' directive and proper component wrapping
- Resolves Functions cannot be passed to Client Components error
